pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mlt-backend"
version = "0.1.0"
description = "Military Leaders Tool backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy[asyncio]>=2.0.32",
  "aiosqlite>=0.20.0",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.4.0",
  "pyjwt>=2.9.0",
  "argon2-cffi>=23.1.0",
  "email-validator>=2.2.0",
  "python-multipart>=0.0.9",
]

[project.optional-dependencies]
test = [
  "pytest>=8.3.2",
  "pytest-asyncio>=0.23.8",
  "httpx>=0.27.0",
  "asgi-lifespan>=2.1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
